Integration tx: move out fee expects
