+++ test/integration/tx/ethereum.js
@@ -2,22 +2,13 @@
 /* eslint-disable no-unused-expressions */
 import chai, { expect } from 'chai'
 import chaiAsPromised from 'chai-as-promised'
-import * as bitcoin from 'bitcoinjs-lib'
-import { hash160 } from '../../../packages/crypto/lib'
-import { calculateFee } from '../../../packages/bitcoin-utils/lib'
-import { addressToString } from '../../../packages/utils/lib'
-import { chains, importBitcoinAddresses, getNewAddress, getRandomBitcoinAddress, mineBlock, fundWallet, describeExternal, connectMetaMask, deployERC20Token, stopEthAutoMining, sleep } from '../common'
+import { chains, getNewAddress, mineBlock, fundWallet, describeExternal, connectMetaMask, deployERC20Token, stopEthAutoMining, expectEthereumFee } from '../common'
 import config from '../config'
 
 process.env['NODE_TLS_REJECT_UNAUTHORIZED'] = 0
 
 chai.use(chaiAsPromised)
 chai.use(require('chai-bignumber')())
-
-async function expectFee (chain, txHash, gasPrice) {
-  const tx = await chain.client.chain.getTransactionByHash(txHash)
-  expect(parseInt(tx._raw.gasPrice, 16)).to.equal(gasPrice)
-}
 
 function testTransaction (chain) {
   it('Sent value to 1 address', async () => {
@@ -43,7 +34,7 @@ function testTransaction (chain) {
     const balAfter = await chain.client.chain.getBalance(addr)
 
     expect(balBefore.plus(value).toString()).to.equal(balAfter.toString())
-    await expectFee(chain, txHash, 10)
+    await expectEthereumFee(chain, txHash, 10)
   })
 
   it('Update transaction fee', async () => {
@@ -52,11 +43,11 @@ function testTransaction (chain) {
 
     const balBefore = await chain.client.chain.getBalance(addr)
     const txHash = await chain.client.chain.sendTransaction(addr, value, undefined, 10)
-    await expectFee(chain, txHash, 10)
+    await expectEthereumFee(chain, txHash, 10)
 
     const newTxHash = await chain.client.chain.updateTransactionFee(txHash, 20)
     expect(newTxHash).to.not.equal(txHash)
-    await expectFee(chain, newTxHash, 20)
+    await expectEthereumFee(chain, newTxHash, 20)
     await mineBlock(chain)
 
     const balAfter = await chain.client.chain.getBalance(addr)
@@ -65,7 +56,7 @@ function testTransaction (chain) {
   })
 }
 
-describe.only('Transactions', function () {
+describe('Transactions', function () {
   this.timeout(config.timeout)
 
   stopEthAutoMining(chains.ethereumWithNode)
